+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 <!-- For changes in existing functionality -->
 
+- Improved the error message for missing highlight configuration errors.
+
 ### Deprecated
 <!-- For soon-to-be removed features -->